commit refactored files session 4

+++ lib/src/models/OrderItemModel.dart
@@ -48,7 +48,7 @@ class OrderItemModel {
    }
 
    Map toJson () => {
-      "id" : (id as int),
+      "id" : id,
       "name" : name,
       "pic" : pic,
       "details" : details,
